[RfC] Set the default navbar state to 'plain'
This change is to make the navbar on the Wikipedia Dashboard to expand automatically. Currently, the behavior is to *hide* the navbar when the dashboard grouping is expanded. That is, two clicks are required to expand the RfC listing rather than one. It can be overridden by setting the `state` parameter.